requirements.txt
Flask==2.0.1
Pillow==8.3.1
datetime==4.3
Werkzeug==2.0.1